+++ app/Http/Controllers/Tasks/TasksController.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\Controller;
 use App\Models\Task;
+use App\Tools\Misc\UserSettings;
 use App\Tools\Webhooks\Webhooks;
 use Auth;
 use Illuminate\Http\Request;
@@ -17,36 +18,42 @@ class TasksController extends Controller {
     }
 
     public function createTask(Request $request) {
-        $title = $request->input('title');
-        $public = $request->input('public');
+        if (UserSettings::hasNode(auth()->user(), UserSettings::CREATE_TASK)) {
+            $title = $request->input('title');
+            $public = $request->input('public');
 
-        if (!$title)
-            $title = 'Untitled';
+            if (!$title)
+                $title = 'Untitled';
 
-        if (!$public)
-            $public = false;
+            if (!$public)
+                $public = false;
 
-        $assignee = $request->input('assignee_id');
+            $assignee = $request->input('assignee_id');
 
-        $task = new Task();
-        $task->title = $title;
-        $task->user_id = Auth::user()->id;
-        $task->assignee_id = $assignee;
-        $task->public = $public;
-        $task->content = $request->input('content');
-        $task->save();
+            $task = new Task();
+            $task->title = $title;
+            $task->user_id = Auth::user()->id;
+            $task->assignee_id = $assignee;
+            $task->public = $public;
+            $task->content = $request->input('content');
+            $task->save();
 
-        return redirect()->back();
+            return redirect()->back();
+        }
     }
 
     public function deleteTask($id) {
-        Task::find($id)->delete();
-        return redirect()->back();
+        if (UserSettings::hasNode(auth()->user(), UserSettings::DELETE_TASK)) {
+            Task::find($id)->delete();
+            return redirect()->back();
+        }
     }
 
     public function completeTask($id) {
-        Task::find($id)->update(['status' => 1]);
-        return redirect()->back();
+        if (UserSettings::hasNode(auth()->user(), UserSettings::MODIFY_TASK)) {
+            Task::find($id)->update(['status' => 1]);
+            return redirect()->back();
+        }
     }
 
     public function refreshIssues() {
